Refactor fetcher to centralized API builder
The README.md file was updated to reflect the changes in the Vanilla JS Fetcher library. The library now includes an API builder which centralizes API configurations and increases code maintainability and reusability. The use of request objects was replaced with the new Api object, which is more flexible and convenient to use. Some code examples were also modified to demonstrate the use of the new features.