"Remove 'isCached' option from API
The 'isCached' option was removed from the API, fetcher, and the relevant tests due to redundancy. The caching of API requests is now solely controlled by the 'cacheInterval' option. If 'cacheInterval' is undefined or less than or equal to 0, the request will not be cached. This simplifies controlling the caching behavior of the requests."