Upgrade @ethang/util to version 1.7.2 and refactor request configuration
This commit updates the version of @ethang/util dependency to "^1.7.2" from "^1.5.0". The 'deepMerge' method from this updated utility package is now being used to refactor the request configuration in 'src/api.ts'. This allows for a more streamlined merging of configuration objects, improving code readability and maintainability. The 'pnpm-lock.yaml' file was updated in the process to reflect this version change.