Update dependencies versions in 'pnpm-lock.yaml' and 'package.json'
This commit updates the versions of several dependencies to their latest stable versions. Changes were made in 'pnpm-lock.yaml' and 'package.json'. Specifically, '@ethang/util' has been updated for bug fixes and feature improvements, '@types/node' was updated to pull in latest type definitions, and 'eslint-config-ethang' was updated to apply the latest set of ESLint rules. This is necessary to keep the project up-to-date with the latest package versions to ensure compatibility, stability and to incorporate new features or improvements.