Add linting to post-version scripts and adjust API test import
Added 'LINT' to 'postVersionBumpScripts' in 'build.mjs' to ensure code style is consistent and problems are detected early. Also, adjusted the import in 'api.test.ts' from 'dist/api' to 'src/api' to test the most recent code in the API directly from the source, which serves for faster and more accurate testing.